Yr/BadeTemp: Rename of plugin and attempted language support
