--- BadeTemp/README.md
@@ -1,1 +1,0 @@
-Viser badetemperatur for noen områder i Oslo
