+++ Badetemperatur/config.py
@@ -28,41 +28,22 @@
 
 ###
 
-"""
-BadeTemp: Viser badetemperatur for steder rundt om i Norge, fra yr.no
-"""
-
-import supybot
-import supybot.world as world
-
-# Use this for the version of this plugin.  You may wish to put a CVS keyword
-# in here if you're keeping the plugin in CVS or some similar system.
-__version__ = ""
-
-# XXX Replace this with an appropriate author or supybot.Author instance.
-__author__ = supybot.authors.unknown
-
-# This is a dictionary mapping supybot.Author instances to lists of
-# contributions.
-__contributors__ = {}
-
-# This is a url where the most recent plugin package can be downloaded.
-__url__ = ''
-
-from . import config
-from . import plugin
-from importlib import reload
-# In case we're being reloaded.
-reload(config)
-reload(plugin)
-# Add more reloads here if you add third-party modules and want them to be
-# reloaded when this plugin is reloaded.  Don't forget to import them as well!
-
-if world.testing:
-    from . import test
-
-Class = plugin.Class
-configure = config.configure
+import supybot.conf as conf
+import supybot.registry as registry
+try:
+    from supybot.i18n import PluginInternationalization
+    _ = PluginInternationalization('Badetemperatur')
+except ImportError:
+    _ = lambda x: x
 
 
-# vim:set shiftwidth=4 tabstop=4 expandtab textwidth=79:
+def configure(advanced):
+    from supybot.questions import expect, anything, something, yn
+    conf.registerPlugin('Badetemperatur', True)
+
+
+Badetemperatur = conf.registerPlugin('Badetemperatur')
+conf.registerGlobalValue(Badetemperatur, 'language', registry.String('', """Override
+    the language for this plugin. Leave empty to use the global
+    supybot.language setting. Accepts any locale code with a matching .po
+    file, including 'no' (Norwegian)."""))
