+++ Badetemperatur/locales/no.po
@@ -5,7 +5,7 @@ msgstr ""
 "Content-Type: text/plain; charset=UTF-8\n"
 
 #: plugin.py
-msgid "<sted>\n\nViser badetemperatur for plasser rundt om i Norge. Data hentes fra Yr.no."
+msgid "<location>\n\nShows bathing water temperatures for locations around Norway. Data is sourced from Yr.no and is Norway-only."
 msgstr "<sted>\n\nViser badetemperatur for plasser rundt om i Norge. Data hentes fra Yr.no."
 
 #: plugin.py
