+++ Yr/__init__.py
@@ -1,4 +1,3 @@
-# coding=utf8
 ###
 # Copyright (c) 2010, Terje Hoås
 # All rights reserved.
@@ -30,45 +29,25 @@
 ###
 
 """
-Add a description of the plugin (to be presented to the user inside the wizard)
-here.  This should describe *what* the plugin does.
+Yr: Fetches weather information from the met.no and yr.no APIs.
 """
 
-import sys
 import supybot
-from supybot import world
+import supybot.world as world
 
-# Use this for the version of this plugin.  You may wish to put a CVS keyword
-# in here if you're keeping the plugin in CVS or some similar system.
 __version__ = ""
-
-# XXX Replace this with an appropriate author or supybot.Author instance.
 __author__ = supybot.Author('Terje Hoås', 'Hoaas', 'supybot@robogoat.dev')
-
-# This is a dictionary mapping supybot.Author instances to lists of
-# contributions.
 __contributors__ = {}
-
-# This is a url where the most recent plugin package can be downloaded.
 __url__ = ''
 
 from . import config
 from . import plugin
-if sys.version_info >= (3, 4):
-    from importlib import reload
-else:
-    from importlib import reload
-# In case we're being reloaded.
+from importlib import reload
 reload(config)
 reload(plugin)
-# Add more reloads here if you add third-party modules and want them to be
-# reloaded when this plugin is reloaded.  Don't forget to import them as well!
 
 if world.testing:
     from . import test
 
 Class = plugin.Class
 configure = config.configure
-
-
-# vim:set shiftwidth=4 tabstop=4 expandtab textwidth=79:
